requirement.txt
langchain
langgraph
python-dotenv
chromadb
openai